feat: stage-aware song routing, unified dev launcher, agent run screen
Songs browser is now the app root (/); /songs redirects there.
Song cards show Ideation/Generation/Composition stage badges and
route directly to the right screen on click. Generate workflow
moves to a dedicated /agent page. Backend scan_songs gains a
`stage` field derived from song_context.yml + composition.yml.
white_work_flow.sh launches API + Next.js with a single command.
